seperate server file
