+++ server/controllers/generate.py
