--- server/csck2notice.json
@@ -1,1 +1,0 @@
-[{"file_name" : "csck2notice.py","inner_idx" : "112","title": "[2017 특성화 우수 학부생 장학금] 규정 변경 안내(중요)","link":"http://csck2.hanyang.ac.kr//front/community/notice/notice-view?id=112","datetime":"2017-02-16"},{"file_name" : "csck2notice.py","inner_idx" : "45","title": "소프트웨어 장기 해외 인턴십 참가자 지원 안내														수정됨","link":"http://csck2.hanyang.ac.kr//front/community/notice/view?id=45","datetime":"2017-05-29"},{"file_name" : "csck2notice.py","inner_idx" : "120","title": "[특성화사업단] 설문조사참여 안내(커피쿠폰 선착순 지급)","link":"http://csck2.hanyang.ac.kr//front/community/notice/view?id=120","datetime":"2017-05-24"},{"file_name" : "csck2notice.py","inner_idx" : "119","title": "[특성화사업단] 학부생 멘토 선정 기준 안내","link":"http://csck2.hanyang.ac.kr//front/community/notice/view?id=119","datetime":"2017-05-23"},{"file_name" : "csck2notice.py","inner_idx" : "94","title": "[2017] 컴퓨터공학부 동아리 활동 지원 안내","link":"http://csck2.hanyang.ac.kr//front/community/notice/view?id=94","datetime":"2016-11-23"},{"file_name" : "csck2notice.py","inner_idx" : "101","title": "소프트웨어 졸업프로젝트 재료/장비/기자재 지원","link":"http://csck2.hanyang.ac.kr//front/community/notice/view?id=101","datetime":"2016-12-09"},{"file_name" : "csck2notice.py","inner_idx" : "117","title": "국내외 소프트웨어 경진/경시대회 참가자 지원안내","link":"http://csck2.hanyang.ac.kr//front/community/notice/view?id=117","datetime":"2017-03-09"}]
