--- server/csgradu.json
@@ -1,1 +1,0 @@
-[{"file_name" : "csgradu.py","inner_idx" : "26813","title": "2017년 1차평가 결과 및 2차평가 일정 공지-장소 추가","link":"http://cs.hanyang.ac.kr/board/gradu_board.php?ptype=view&idx=26813&page=&code=gradu_board","datetime":"17.05.18"},{"file_name" : "csgradu.py","inner_idx" : "26626","title": "2017년 졸업프로젝트 제안서 양식 (4/5일 수정)","link":"http://cs.hanyang.ac.kr/board/gradu_board.php?ptype=view&idx=26626&page=&code=gradu_board","datetime":"17.03.17"},{"file_name" : "csgradu.py","inner_idx" : "26543","title": "2017년 졸업 프로젝트수강신청 배정 추가 공지(계속 업데이트)","link":"http://cs.hanyang.ac.kr/board/gradu_board.php?ptype=view&idx=26543&page=&code=gradu_board","datetime":"17.02.08"},{"file_name" : "csgradu.py","inner_idx" : "26535","title": "2017년 졸업 프로젝트 지도교수 및 수강신청 배정 공지_170308수정","link":"http://cs.hanyang.ac.kr/board/gradu_board.php?ptype=view&idx=26535&page=&code=gradu_board","datetime":"17.02.06"},{"file_name" : "csgradu.py","inner_idx" : "26503","title": "2017년 지정주제(지도 교수 선정 관련) 추가 공지입니다.","link":"http://cs.hanyang.ac.kr/board/gradu_board.php?ptype=view&idx=26503&page=&code=gradu_board","datetime":"17.01.12"},{"file_name" : "csgradu.py","inner_idx" : "26411","title": "2017년 지정주제 공지입니다.","link":"http://cs.hanyang.ac.kr/board/gradu_board.php?ptype=view&idx=26411&page=&code=gradu_board","datetime":"16.12.12"},{"file_name" : "csgradu.py","inner_idx" : "26409","title": "2017 졸업프로젝트 설명회 ppt 자료입니다.","link":"http://cs.hanyang.ac.kr/board/gradu_board.php?ptype=view&idx=26409&page=&code=gradu_board","datetime":"16.12.09"},{"file_name" : "csgradu.py","inner_idx" : "20260","title": "2016년 졸업프로젝트 기자재 관련 공지","link":"http://cs.hanyang.ac.kr/board/gradu_board.php?ptype=view&idx=20260&page=&code=gradu_board","datetime":"15.12.30"},{"file_name" : "csgradu.py","inner_idx" : "20259","title": "자주 묻는 질문들(자유/지정주제, 휴학 중 졸프진행)","link":"http://cs.hanyang.ac.kr/board/gradu_board.php?ptype=view&idx=20259&page=&code=gradu_board","datetime":"15.12.29"},{"file_name" : "csgradu.py","inner_idx" : "26723","title": "졸업프로젝트 산학협력지원 프로그램 설명회 (4월 17일 11:30 - 13:00 / ITBI 508호)","link":"http://cs.hanyang.ac.kr/board/gradu_board.php?ptype=view&idx=26723&page=1&code=gradu_board","datetime":"17.04.13"},{"file_name" : "csgradu.py","inner_idx" : "26504","title": "2017년 졸업 프로젝트 조교가 변경되었습니다.","link":"http://cs.hanyang.ac.kr/board/gradu_board.php?ptype=view&idx=26504&page=1&code=gradu_board","datetime":"17.01.12"},{"file_name" : "csgradu.py","inner_idx" : "26369","title": "졸업프로젝트 설명회 일정 공지","link":"http://cs.hanyang.ac.kr/board/gradu_board.php?ptype=view&idx=26369&page=1&code=gradu_board","datetime":"16.11.25"},{"file_name" : "csgradu.py","inner_idx" : "26352","title": "2차 평가 합격자 공지","link":"http://cs.hanyang.ac.kr/board/gradu_board.php?ptype=view&idx=26352&page=1&code=gradu_board","datetime":"16.11.25"},{"file_name" : "csgradu.py","inner_idx" : "20277","title": "2차평가 순서 안내.","link":"http://cs.hanyang.ac.kr/board/gradu_board.php?ptype=view&idx=20277&page=1&code=gradu_board","datetime":"16.11.15"},{"file_name" : "csgradu.py","inner_idx" : "20276","title": "1차 평가 성적 공지","link":"http://cs.hanyang.ac.kr/board/gradu_board.php?ptype=view&idx=20276&page=1&code=gradu_board","datetime":"16.10.19"},{"file_name" : "csgradu.py","inner_idx" : "20275","title": "1차 & 2차 평가 일정 안내","link":"http://cs.hanyang.ac.kr/board/gradu_board.php?ptype=view&idx=20275&page=1&code=gradu_board","datetime":"16.09.30"},{"file_name" : "csgradu.py","inner_idx" : "20274","title": "최종 결과 보고서 파일 및 일정(일정 재공지)","link":"http://cs.hanyang.ac.kr/board/gradu_board.php?ptype=view&idx=20274&page=1&code=gradu_board","datetime":"16.09.30"},{"file_name" : "csgradu.py","inner_idx" : "20273","title": "졸업프로젝트 설명회 ppt자료입니다.","link":"http://cs.hanyang.ac.kr/board/gradu_board.php?ptype=view&idx=20273&page=1&code=gradu_board","datetime":"16.09.05"},{"file_name" : "csgradu.py","inner_idx" : "20272","title": "2016년도 2학기 졸업프로젝트 설명회 안내.","link":"http://cs.hanyang.ac.kr/board/gradu_board.php?ptype=view&idx=20272&page=1&code=gradu_board","datetime":"16.08.19"},{"file_name" : "csgradu.py","inner_idx" : "20271","title": "2016년도 2학기 지정주제 공지입니다.","link":"http://cs.hanyang.ac.kr/board/gradu_board.php?ptype=view&idx=20271&page=1&code=gradu_board","datetime":"16.08.05"},{"file_name" : "csgradu.py","inner_idx" : "20270","title": "2016년도 2학기 제안서 입니다.(양식 첨부)","link":"http://cs.hanyang.ac.kr/board/gradu_board.php?ptype=view&idx=20270&page=1&code=gradu_board","datetime":"16.08.05"},{"file_name" : "csgradu.py","inner_idx" : "20269","title": "주제 변경 사유서입니다.","link":"http://cs.hanyang.ac.kr/board/gradu_board.php?ptype=view&idx=20269&page=1&code=gradu_board","datetime":"16.07.18"},{"file_name" : "csgradu.py","inner_idx" : "20268","title": "2차 평가 통과자 발표","link":"http://cs.hanyang.ac.kr/board/gradu_board.php?ptype=view&idx=20268&page=1&code=gradu_board","datetime":"16.06.03"},{"file_name" : "csgradu.py","inner_idx" : "20267","title": "중간보고서 제출 공지","link":"http://cs.hanyang.ac.kr/board/gradu_board.php?ptype=view&idx=20267&page=1&code=gradu_board","datetime":"16.05.18"},{"file_name" : "csgradu.py","inner_idx" : "20266","title": "1차 평가 결과 & 2차 평가일 공지.","link":"http://cs.hanyang.ac.kr/board/gradu_board.php?ptype=view&idx=20266&page=1&code=gradu_board","datetime":"16.04.29"},{"file_name" : "csgradu.py","inner_idx" : "20265","title": "2016년도 1학기 1,2차 평가 안내(필독)","link":"http://cs.hanyang.ac.kr/board/gradu_board.php?ptype=view&idx=20265&page=1&code=gradu_board","datetime":"16.03.30"},{"file_name" : "csgradu.py","inner_idx" : "20264","title": "2차 제안서 제출일 공지.","link":"http://cs.hanyang.ac.kr/board/gradu_board.php?ptype=view&idx=20264&page=1&code=gradu_board","datetime":"16.02.03"},{"file_name" : "csgradu.py","inner_idx" : "20263","title": "2016년도 1학기 수강신청 배정입니다.(필독)","link":"http://cs.hanyang.ac.kr/board/gradu_board.php?ptype=view&idx=20263&page=1&code=gradu_board","datetime":"16.02.03"},{"file_name" : "csgradu.py","inner_idx" : "20262","title": "2016년도 1학기 졸프 시작학생들을 위한 제안서 공지입니다.","link":"http://cs.hanyang.ac.kr/board/gradu_board.php?ptype=view&idx=20262&page=1&code=gradu_board","datetime":"15.12.31"}]
