--- server/csjob.json
@@ -1,1 +1,0 @@
-[{"file_name" : "csjob.py","inner_idx" : "26431","title": "[SW중심대학] 동계인턴십 추천결과 안내","link":"http://cs.hanyang.ac.kr/board/job_board.php?ptype=view&idx=26431&page=&code=job_board","datetime":"16.12.15"},{"file_name" : "csjob.py","inner_idx" : "26830","title": "[오비맥주] 2017 AB InBev Korea(오비맥주) GMT Summer Intern 모집","link":"http://cs.hanyang.ac.kr/board/job_board.php?ptype=view&idx=26830&page=1&code=job_board","datetime":"17.05.25"},{"file_name" : "csjob.py","inner_idx" : "26829","title": "[UMZi] 모바일 쇼핑의 패러다임을 바꾸고 싶은 인력 채용","link":"http://cs.hanyang.ac.kr/board/job_board.php?ptype=view&idx=26829&page=1&code=job_board","datetime":"17.05.25"},{"file_name" : "csjob.py","inner_idx" : "26828","title": "2017년 서울우유 5급 신입 및 경력직원 공개채용","link":"http://cs.hanyang.ac.kr/board/job_board.php?ptype=view&idx=26828&page=1&code=job_board","datetime":"17.05.24"},{"file_name" : "csjob.py","inner_idx" : "26827","title": "KB국민카드 - 2017년 KB국민카드 IT직무 신입사원 채용 공고(~5/22)","link":"http://cs.hanyang.ac.kr/board/job_board.php?ptype=view&idx=26827&page=1&code=job_board","datetime":"17.05.24"},{"file_name" : "csjob.py","inner_idx" : "26826","title": "욕심있는 예비 개발자 SW교육프로그램 '부스트캠프' 모집 안내","link":"http://cs.hanyang.ac.kr/board/job_board.php?ptype=view&idx=26826&page=1&code=job_board","datetime":"17.05.24"},{"file_name" : "csjob.py","inner_idx" : "26825","title": "[마이다스아이티]2017 K-SW포럼 인턴십 모집홍보","link":"http://cs.hanyang.ac.kr/board/job_board.php?ptype=view&idx=26825&page=1&code=job_board","datetime":"17.05.24"},{"file_name" : "csjob.py","inner_idx" : "26824","title": "(미래창조과학부) 이공계전문기술연수사업 연수생 모집 사업 안내","link":"http://cs.hanyang.ac.kr/board/job_board.php?ptype=view&idx=26824&page=1&code=job_board","datetime":"17.05.24"},{"file_name" : "csjob.py","inner_idx" : "26823","title": "[LS엠트론]2차 2017 상반기 채용연계형 하계 인턴 및 산학장학생 선발 안내","link":"http://cs.hanyang.ac.kr/board/job_board.php?ptype=view&idx=26823&page=1&code=job_board","datetime":"17.05.24"},{"file_name" : "csjob.py","inner_idx" : "26822","title": "[이수그룹] 2017 이수그룹 채용연계형 인턴십 ( ~ 6/14 (수) 17시 까지 접수)","link":"http://cs.hanyang.ac.kr/board/job_board.php?ptype=view&idx=26822&page=1&code=job_board","datetime":"17.05.24"},{"file_name" : "csjob.py","inner_idx" : "26821","title": "[(주)한솥] 2017년 신입사원 공채 안내","link":"http://cs.hanyang.ac.kr/board/job_board.php?ptype=view&idx=26821&page=1&code=job_board","datetime":"17.05.24"},{"file_name" : "csjob.py","inner_idx" : "26814","title": "디지털 미디어 연구실 인턴 연구원 모집 안내","link":"http://cs.hanyang.ac.kr/board/job_board.php?ptype=view&idx=26814&page=1&code=job_board","datetime":"17.05.18"},{"file_name" : "csjob.py","inner_idx" : "26810","title": "LG Display Future Dream Idea 대학(원)생 IDEA 공모전","link":"http://cs.hanyang.ac.kr/board/job_board.php?ptype=view&idx=26810&page=1&code=job_board","datetime":"17.05.17"},{"file_name" : "csjob.py","inner_idx" : "26809","title": "게임빌컴투스플랫폼 부문 채용(신입)","link":"http://cs.hanyang.ac.kr/board/job_board.php?ptype=view&idx=26809&page=1&code=job_board","datetime":"17.05.17"},{"file_name" : "csjob.py","inner_idx" : "26808","title": "[마이다스아이티]K-SW 포럼 인턴십 모집","link":"http://cs.hanyang.ac.kr/board/job_board.php?ptype=view&idx=26808&page=1&code=job_board","datetime":"17.05.17"},{"file_name" : "csjob.py","inner_idx" : "26807","title": "[LS엠트론] 2017 상반기 채용연계형 하계 인턴 및 산학장학생 선발 안내","link":"http://cs.hanyang.ac.kr/board/job_board.php?ptype=view&idx=26807&page=1&code=job_board","datetime":"17.05.17"},{"file_name" : "csjob.py","inner_idx" : "26806","title": "[NTS] 2017년 하계 인턴 모집","link":"http://cs.hanyang.ac.kr/board/job_board.php?ptype=view&idx=26806&page=1&code=job_board","datetime":"17.05.17"},{"file_name" : "csjob.py","inner_idx" : "26805","title": "[네오비즈] 앱 및 웹 개발 가능한 학생 모집","link":"http://cs.hanyang.ac.kr/board/job_board.php?ptype=view&idx=26805&page=1&code=job_board","datetime":"17.05.17"},{"file_name" : "csjob.py","inner_idx" : "26804","title": "[우리에프아이에스]2017년 신입사원 채용 안내","link":"http://cs.hanyang.ac.kr/board/job_board.php?ptype=view&idx=26804&page=1&code=job_board","datetime":"17.05.17"},{"file_name" : "csjob.py","inner_idx" : "26803","title": "레드덕] 전문연구요원 채용 공고 안내","link":"http://cs.hanyang.ac.kr/board/job_board.php?ptype=view&idx=26803&page=1&code=job_board","datetime":"17.05.17"},{"file_name" : "csjob.py","inner_idx" : "26796","title": "삼성전자 DS부문 Tech Day 안내","link":"http://cs.hanyang.ac.kr/board/job_board.php?ptype=view&idx=26796&page=1&code=job_board","datetime":"17.05.15"}]
