--- server/csstrk.json
@@ -1,1 +1,0 @@
-[{"file_name" : "csstrk.py","inner_idx" : "26787","title": "[설명회/지원]2017-1학기 삼성 SCSC 인력양성 프로그램 설명회(6/2) 및 지원(~7/31) 안내","link":"http://cs.hanyang.ac.kr/board/trk_board.php?ptype=view&idx=26787&page=&code=trk_board","datetime":"17.05.11"},{"file_name" : "csstrk.py","inner_idx" : "26774","title": "[계절수업]2017학년도 삼성SCSC프로그램 여름 계절학기 수강신청 안내(~5/30)","link":"http://cs.hanyang.ac.kr/board/trk_board.php?ptype=view&idx=26774&page=&code=trk_board","datetime":"17.05.02"},{"file_name" : "csstrk.py","inner_idx" : "26745","title": "[대체과목신청] 2017학년도 삼성 SCSC 트랙 과목 대체 승인 신청 안내(~5/12)","link":"http://cs.hanyang.ac.kr/board/trk_board.php?ptype=view&idx=26745&page=&code=trk_board","datetime":"17.04.21"},{"file_name" : "csstrk.py","inner_idx" : "12533","title": "[수업개설 일정] 삼성 SCSC 트랙 개설 일정 안내","link":"http://cs.hanyang.ac.kr/board/trk_board.php?ptype=view&idx=12533&page=&code=trk_board","datetime":"16.05.31"},{"file_name" : "csstrk.py","inner_idx" : "12523","title": "[교과목] 삼성트랙 지정 교과목 변경 안내","link":"http://cs.hanyang.ac.kr/board/trk_board.php?ptype=view&idx=12523&page=&code=trk_board","datetime":"16.01.18"},{"file_name" : "csstrk.py","inner_idx" : "26603","title": "[2017년 상반기 삼성전자 3급 신입채용] 지원 안내","link":"http://cs.hanyang.ac.kr/board/trk_board.php?ptype=view&idx=26603&page=1&code=trk_board","datetime":"17.03.10"},{"file_name" : "csstrk.py","inner_idx" : "26600","title": "[멘토링]2017-1학기 SCSC 멘토링 팀배정 안내","link":"http://cs.hanyang.ac.kr/board/trk_board.php?ptype=view&idx=26600&page=1&code=trk_board","datetime":"17.03.09"},{"file_name" : "csstrk.py","inner_idx" : "26558","title": "[멘토링] 삼성 SCSC 비전공 트랙 참여학생 멘토링 지원 안내(SCSC과정 학생)","link":"http://cs.hanyang.ac.kr/board/trk_board.php?ptype=view&idx=26558&page=1&code=trk_board","datetime":"17.02.20"},{"file_name" : "csstrk.py","inner_idx" : "26549","title": "[선발명단] 삼성 SCSC 선발 명단(2017-1학기)","link":"http://cs.hanyang.ac.kr/board/trk_board.php?ptype=view&idx=26549&page=1&code=trk_board","datetime":"17.02.08"},{"file_name" : "csstrk.py","inner_idx" : "26507","title": "[수강신청] 2017-1학기 삼성SCSC  수강신청 안내_수강허용확인서 포함(업데이트)","link":"http://cs.hanyang.ac.kr/board/trk_board.php?ptype=view&idx=26507&page=1&code=trk_board","datetime":"17.01.13"},{"file_name" : "csstrk.py","inner_idx" : "26412","title": "[지원] 2017-1학기 삼성 SCSC 지원 안내(~2/3)","link":"http://cs.hanyang.ac.kr/board/trk_board.php?ptype=view&idx=26412&page=1&code=trk_board","datetime":"16.12.12"},{"file_name" : "csstrk.py","inner_idx" : "12539","title": "[선발명단] 삼성 SCSC 추가 선발 명단(2016-2학기)","link":"http://cs.hanyang.ac.kr/board/trk_board.php?ptype=view&idx=12539&page=1&code=trk_board","datetime":"16.09.08"},{"file_name" : "csstrk.py","inner_idx" : "12538","title": "[멘토링] 16-2학기 SCSC 멘토링 팀배정 안내","link":"http://cs.hanyang.ac.kr/board/trk_board.php?ptype=view&idx=12538&page=1&code=trk_board","datetime":"16.09.06"},{"file_name" : "csstrk.py","inner_idx" : "12537","title": "[설명회] 2016-2학기 삼성 SCSC 설명회 및 지원 안내","link":"http://cs.hanyang.ac.kr/board/trk_board.php?ptype=view&idx=12537&page=1&code=trk_board","datetime":"16.08.23"},{"file_name" : "csstrk.py","inner_idx" : "12536","title": "[선발명단] 삼성 SCSC 선발 명단(2016-2학기)","link":"http://cs.hanyang.ac.kr/board/trk_board.php?ptype=view&idx=12536&page=1&code=trk_board","datetime":"16.08.19"},{"file_name" : "csstrk.py","inner_idx" : "12535","title": "[멘토링] 삼성 SCSC 비전공 트랙 참여학생 멘토링 지원 안내(SCSC과정 학생)","link":"http://cs.hanyang.ac.kr/board/trk_board.php?ptype=view&idx=12535&page=1&code=trk_board","datetime":"16.08.18"},{"file_name" : "csstrk.py","inner_idx" : "12534","title": "[수강신청] 2016-2학기 삼성scsc 수강신청 안내(첨부_수강허용확인서)","link":"http://cs.hanyang.ac.kr/board/trk_board.php?ptype=view&idx=12534&page=1&code=trk_board","datetime":"16.07.26"},{"file_name" : "csstrk.py","inner_idx" : "12532","title": "[설명회] 2016-1학기 삼성 SCSC 설명회 및 지원 안내","link":"http://cs.hanyang.ac.kr/board/trk_board.php?ptype=view&idx=12532&page=1&code=trk_board","datetime":"16.05.31"},{"file_name" : "csstrk.py","inner_idx" : "12531","title": "[대체과목승인] 2016-1학기 SCSC 트랙 과목 대체 승인 신청 결과 안내","link":"http://cs.hanyang.ac.kr/board/trk_board.php?ptype=view&idx=12531&page=1&code=trk_board","datetime":"16.05.18"},{"file_name" : "csstrk.py","inner_idx" : "12530","title": "[계절수업] 삼성SCSC 여름계절학기 과목 개설 안내","link":"http://cs.hanyang.ac.kr/board/trk_board.php?ptype=view&idx=12530&page=1&code=trk_board","datetime":"16.04.07"},{"file_name" : "csstrk.py","inner_idx" : "12529","title": "[대체과목신청] 2016-1학기 SCSC 트랙 과목 대체 승인 신청 안내","link":"http://cs.hanyang.ac.kr/board/trk_board.php?ptype=view&idx=12529&page=1&code=trk_board","datetime":"16.04.06"},{"file_name" : "csstrk.py","inner_idx" : "12528","title": "[멘토링] 삼성 SCSC 비전공 트랙 참여학생 멘토링 지원 안내(SCSC과정 학생)","link":"http://cs.hanyang.ac.kr/board/trk_board.php?ptype=view&idx=12528&page=1&code=trk_board","datetime":"16.02.19"},{"file_name" : "csstrk.py","inner_idx" : "12527","title": "[선발명단] 삼성 SCSC 추가 선발 명단(2016-1학기)","link":"http://cs.hanyang.ac.kr/board/trk_board.php?ptype=view&idx=12527&page=1&code=trk_board","datetime":"16.02.15"},{"file_name" : "csstrk.py","inner_idx" : "12526","title": "[수강신청] 2016-1학기 삼성scsc 수강신청 안내(첨부_수강허용확인서)","link":"http://cs.hanyang.ac.kr/board/trk_board.php?ptype=view&idx=12526&page=1&code=trk_board","datetime":"16.02.13"},{"file_name" : "csstrk.py","inner_idx" : "12525","title": "[선발명단] 삼성 SCSC 선발 명단(2016-1학기)","link":"http://cs.hanyang.ac.kr/board/trk_board.php?ptype=view&idx=12525&page=1&code=trk_board","datetime":"16.02.13"}]
