--- server/db/HyLionPost-13a0ebe69f3b.json
@@ -1,12 +1,0 @@
-{
-  "type": "service_account",
-  "project_id": "hylionpost",
-  "private_key_id": "13a0ebe69f3be3dc8e96ea191797027cfd0955a2",
-  "private_key": "-----BEGIN PRIVATE KEY-----\nMIIEvgIBADANBgkqhkiG9w0BAQEFAASCBKgwggSkAgEAAoIBAQCWhGcNeHTa1S17\nvJlUH9BGRP1tVeOzZqs/f8O/V29wsEmKPHPer5U8NBMebkVDNER3NKyG11Sf0Ffl\ntaC/C4OtJjj1kEzyc1Vi3n9srcSNwUoRaeei8a1m0WkVI7W+YNcm2dtygCp9Q9rl\nOrrahkNfhv+1/aCKlPIASZhZaC0tU9onC8KHd7NUT+gsbyHFxwMT1nB/8AfyrENA\n4v363srRuBzqGSOLRrIg1u37hqfU88IQyCQg6KqrDZdo2i2rd5de6gw3r8gnOV4S\n7r44AflVeWihQUvmfiUZPL9D+cruObqr/YfSLdYucIYX5U/RNEUf4DV6Eit0yCDV\nCQzkOYoJAgMBAAECggEAAarsoeBzaf/J6diueblDA15NEQauN80DyXeyzWZMeqsH\nW4/dKErZfFHm4tPNsmeMOSZlIXbWDhG+wZGsOuWnozXc1HKxeY+30tk8f/Gib4Ak\nYBJEceCb31tjXsu3HimARMLN7MQ5BEuJy7elfZmWekLahzOM0XBCSKjIFniCJKII\neOeQyHefCJEjvGoWXwTcobEmy4rmym9mxcFjv3Mf/T2WZ23OClwzNVhoeFYbqaPX\ntzyjbrtT/1uLZCVhwKCS8+FuxUfQ2QCopiE++QwtDIsP90IeVKjq6EYw7pPXolyD\njcuSb0C0PnGIgTcu3IHYcb6CBA9/P1ek/VaqoTfscQKBgQDST34Nu9eJOBJVeiZO\n1qIwgq7Haax5TlYe9iSDqywrbBPVPGN7cKxpRcK1H8e+q/4tK0QQ7FrCdrfyZ6Sp\n369zMxD6Z5gI3HIHo7LzEqz+8NESNirKDQwIvVHHsLPnREqDTVKg2I83WbzSRhU7\nl5hqdjFN2CLQwzJphXxeqckJ+QKBgQC3N3rHP70S6E9FM//VDlnfjG2D6XbH25N1\nGrH1+6hrh9AIvc86UASwOhiP2odAVIrO0emNMQZmh8mXilho3cd/A72xchsjs9aT\ni5t59FF2am98EqeYzLXSbk3f50LRZCoI8ZhbKP1TuniUT27Vn46P7u7vMDsiKj3E\ntr/bQ4sEkQKBgQCv+HuMmkO6OQR+KPqstMJ4vJ/pku4Rgg5lUacz1KHco03k+eZz\nfs+ENi64xqKnU9oViUImtt0qIZPAbCBHUzbGuU/XdZ6sq2duJPaylEnT90O4bUJY\nVZpgJair4ZpmQ6nlO2mL1CfhNGDdwG47spiBW+vdKp/W9bzUtjBeNMpWYQKBgQCB\nWcTv+PsG50ZX6E9dXwVqg10qPWCj10zy50bvEJ4mVLs2ALPWSnoU5N9EaJcGfumE\n/MQz74HCT2mu+HhUiQd4j+6T4ahxe10py+2XmzsMLDbLVOxIP3a6uy91vXL6HApl\nSBQ1f+nh/At3QeFRJpXi3mCIM/f7tfDn+dHX/oaeUQKBgAW6AHEKnII9hC2OVofl\nhN9dxio7+GhcIoZG2P8nbwOh2Ebu0GMGatm83QK7op+9Y3Z22y2lcN1sRAsN0qWN\nv7OYOjaCKFp2D26VmECPAN23Mu9zS8PH1St0rpIKOiSaIAN3UBVLTj3HmjJSHSRp\nNlpZK4SculPtNtkEGiCyodqq\n-----END PRIVATE KEY-----\n",
-  "client_email": "firebase-adminsdk-dbfv4@hylionpost.iam.gserviceaccount.com",
-  "client_id": "115903651930109897671",
-  "auth_uri": "https://accounts.google.com/o/oauth2/auth",
-  "token_uri": "https://accounts.google.com/o/oauth2/token",
-  "auth_provider_x509_cert_url": "https://www.googleapis.com/oauth2/v1/certs",
-  "client_x509_cert_url": "https://www.googleapis.com/robot/v1/metadata/x509/firebase-adminsdk-dbfv4%40hylionpost.iam.gserviceaccount.com"
-}
