+++ server/run.py
@@ -1,0 +1,7 @@
+from header import *
+from views import *
+
+# server running
+
+if __name__ == '__main__':
+    app.run()
